+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v2.0.0 (2023-11-09)
+
+- docker_ps: do not use json format output when executing `docker ps` command
+- docker_ps: Use option --no-trunc when executing `docker ps` command
+- docker_ps: change results key names (ID becomes id, Names becomes name, etc)
+
 ## v1.0.2 (2023-11-07)
 
 - docker_ps: improve doc